fix: change business logic
