+++ src/album/album.module.ts
@@ -1,10 +1,12 @@
 import { Module } from '@nestjs/common';
 import { AlbumService } from './album.service';
 import { AlbumController } from './album.controller';
-import { AlbumRepository } from './album.repository';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { AlbumEntity } from './entity/album.entity';
 
 @Module({
+  imports: [TypeOrmModule.forFeature([AlbumEntity])],
   controllers: [AlbumController],
-  providers: [AlbumService, AlbumRepository],
+  providers: [AlbumService],
 })
 export class AlbumModule {}
